testing

+++ ct/snowshare.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
-source <(curl -fsSL https://raw.githubusercontent.com/community-scripts/ProxmoxVE/main/misc/build.func)
+#source <(curl -fsSL https://raw.githubusercontent.com/community-scripts/ProxmoxVE/main/misc/build.func)
 
-#source <(curl -fsSL https://raw.githubusercontent.com/TuroYT/ProxmoxVED/refs/heads/add-snowshare/misc/build.func)
+source <(curl -fsSL https://raw.githubusercontent.com/TuroYT/ProxmoxVED/refs/heads/add-snowshare/misc/build.func)
 # Copyright (c) 2021-2025 community-scripts ORG
 # Author: TuroYT
 # License: MIT | https://github.com/community-scripts/ProxmoxVE/raw/main/LICENSE
@@ -13,7 +13,7 @@ var_cpu="${var_cpu:-1}"
 var_ram="${var_ram:-1024}"
 var_disk="${var_disk:-5}"
 var_os="${var_os:-debian}"
-var_version="${var_version:-12}"
+var_version="${var_version:-13}"
 var_unprivileged="${var_unprivileged:-1}"
 
 header_info "$APP"
@@ -30,15 +30,24 @@ function update_script() {
     msg_error "No ${APP} Installation Found!"
     exit
   fi
-  msg_info "Updating ${APP}"
-  systemctl stop snowshare
-  cd /opt/snowshare
-  git pull
-  npm ci
-  npx prisma generate
-  npm run build
-  systemctl start snowshare
-  msg_ok "Updated ${APP}"
+
+  if check_for_gh_release "snowshare" "TuroYT/snowshare"; then
+    msg_info "Updating ${APP} to v${RELEASE}"
+
+    systemctl stop snowshare
+    cd /opt/
+    fetch_and_deploy_gh_release "snowshare" "TuroYT/snowshare"
+    cd /opt/snowshare
+    npm ci
+    npx prisma generate
+    npm run build
+    systemctl start snowshare
+    msg_ok "Updated ${APP}"
+    exit
+
+  else
+    msg_ok "No update required. ${APP} is already at v${RELEASE}."
+  fi
   exit
 }
